V0.9.1
- rename software files
- ability to switch operation mode through button push after reset
-
